Fixed bug where tail calls were wrongly assumed.
```
(define (f iterator) (iterator))
(f (lambda () (iterator)))
```
In this scenario the lambda passed to `f` is called with the name
`iterator` due to the naming from the parameter. The lambda therefore
believes it's recursive, since the name it is called with matches whith
the name of the expression in tail position.
This was fixed by evaluating the name for the expression before call if
the name is also among the parameters of the caller.